[fix] 코드 리뷰 내용 적용 - JpaConfiguration 클래스 분리 - 정적 메서드 팩토리 of, from 변경 - 금액 타입 BigDecimal 변경 - 에러 메시지 enum 분리

+++ src/main/java/com/wanted/preonboarding/ticket/controller/model/request/ReservationApplyRequest.java
@@ -2,6 +2,8 @@ package com.wanted.preonboarding.ticket.controller.model.request;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
@@ -10,7 +12,7 @@ import lombok.*;
 public class ReservationApplyRequest {
     private String userName;
     private String phoneNumber;
-    private double amountAvailable;
+    private BigDecimal amountAvailable;
     private int round;
     private Long performanceId;
     private Long performanceSeatId;
